+++ src/primitives.jl
@@ -1,29 +1,29 @@
-invertapply(::typeof(+), t::Type{Tuple{Float64, Float64}}, y, φ) = (v = ℝ(φ); (y - v, v))
-invertapply(::typeof(+), t::Type{Tuple{Int64, Int64}}, y, φ) = (v = ℤ(φ); (y - v, v))
-invertapply(::typeof(+), t::Type{Tuple{T, T}}, y, φ) where T = (v = ℝ(φ); (y - v, v))
+const Float64_2 = Type{Tuple{Float64, Float64}}
 
-invertapply(::typeof(-), t::Type{Tuple{Int64, Int64}}, y, φ) = (v = ℤ(φ); (y + v, v))
-invertapply(::typeof(-), t::Type{Tuple{Float64, Float64}}, y, φ) = (v = ℝ(φ); (y + v, v))
-invertapply(::typeof(-), types::Type{Tuple{T, T}}, y, φ) where T = (v = ℝ(φ); (y + v, v))
+# +, - relation
+choose(θ, ::typeof(+), ::Type{NTuple{2, <:Real}}, ::Type{A}, ::Type{ZB}, z, b) = (z - b,)
+choose(θ, ::typeof(+), ::Type{<:NTuple{2, <:Real}}, ::Type{AB}, ::Type{Z}, z) = 
+  let θ_ = ℝ(θ) ; (z - θ_, θ_) end
+choose(θ, ::typeof(+), ::Type{Tuple{Int, Int}}, ::Type{AB}, ::Type{Z}, z) = 
+  let θ_ = ℝ(θ) ; (z - θ_, θ_) end
 
-function invertapply(::typeof(*), t::Type{Tuple{Float64, Float64}}, y, φ)
-  b = 𝔹(φ[2])
-  v = ℝ(φ[1])
-  b ? (y/v, v) : (v, y/v)
+
+# *, / relation
+choose(θ, ::typeof(*), ::Float64_2, ::Type{B}, ::Type{ZA}, z, a) = 
+  (z / a,)
+
+function choose(θ, ::typeof(*), ::Float64_2, ::Type{AB}, ::Type{Z}, z)
+  b = 𝔹(θ)
+  v = ℝ(θ)
+  b ? (z/v, v) : (v, z/v)
 end
 
-function invertapply(::typeof(*), t::Type{Tuple{Float64, PIConstant{T}}}, constants, y, φ) where T
-  c = constants[1].value
-  (y/c, c)
-end
+choose(θ, ::typeof(/), ::Float64_2, ::Type{AB}, ::Type{Z}, z) =
+  let r = ℝ(θ)
+    (z * r, r)
+  end
 
-function invertapply(::typeof(*), t::Type{Tuple{T, T}}, y, φ) where T
-  b = 𝔹(φ[2])
-  v = ℝ(φ[1])
-  b ? (y/v, v) : (v, y/v)
-end
-invertapply(::typeof(/), types::Type{Tuple{T, T}}, y, φ) where T = (y*ℝ(φ), ℝ(φ))
-
+choose(θ, loc::Loc, args...) = choose(project(θ, loc), args...)
 ### TODO:
 # Handle multiple arguments like :+(%2, %3, %4, %5) (all with potentially different types)
 # Handle inverse integer multiplication and division: factoring problems
